+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+- Bump discord.py to 1.5.0, use Intents
+
 ## [0.2.1]
 
 - `announce` command